requirements.txt
gitdb==4.0.9
GitPython==3.1.29
pick==2.1.0
smmap==5.0.0
tqdm==4.64.1
